Move admin-ui-utils into main package
