+++ .changeset/empty-rockets-destroy.md
@@ -1,0 +1,8 @@
+---
+'@keystone-next/admin-ui-utils': major
+'@keystone-next/auth': major
+'@keystone-next/keystone': minor
+'@keystone-next/fields-document': patch
+---
+
+Moved `@keystone-next/admin-ui-utils` to `@keystone-next/keystone/admin-ui/utils`
