+++ packages/admin-ui-utils/package.json
@@ -4,21 +4,6 @@
   "main": "dist/admin-ui-utils.cjs.js",
   "module": "dist/admin-ui-utils.esm.js",
   "license": "MIT",
-  "dependencies": {
-    "@babel/runtime": "^7.15.3",
-    "@emotion/weak-memoize": "^0.2.5",
-    "@keystone-next/types": "^24.0.0",
-    "@keystone-ui/core": "^3.1.1",
-    "@types/react": "^17.0.18",
-    "fast-deep-equal": "^3.1.3",
-    "graphql": "^15.5.1"
-  },
-  "peerDependencies": {
-    "react": "^17.0.2"
-  },
-  "devDependencies": {
-    "react": "^17.0.2"
-  },
   "engines": {
     "node": "^12.20 || >= 14.13"
   },
