+++ packages/admin-ui-utils/src/index.ts
@@ -1,6 +1,5 @@
-export * from './dataGetter';
-export * from './Fields';
-export * from './getRootGraphQLFieldsFromFieldController';
-export * from './item-form';
-export * from './serialization';
-export * from './useInvalidFields';
+throw new Error(
+  '`@keystone-next/admin-ui-utils` has been moved to `@keystone-next/keystone/admin-ui/utils`, please import from there instead.'
+);
+
+export {};
