+++ packages/auth/package.json
@@ -10,7 +10,6 @@
   "dependencies": {
     "@babel/runtime": "^7.15.3",
     "@graphql-tools/schema": "^8.1.1",
-    "@keystone-next/admin-ui-utils": "^5.0.6",
     "@keystone-next/types": "^24.0.0",
     "@keystone-ui/button": "^5.0.0",
     "@keystone-ui/core": "^3.1.1",
