+++ packages/fields-document/package.json
@@ -22,7 +22,6 @@
     "@babel/runtime": "^7.15.3",
     "@braintree/sanitize-url": "^5.0.2",
     "@emotion/weak-memoize": "^0.2.5",
-    "@keystone-next/admin-ui-utils": "^5.0.6",
     "@keystone-next/keystone": "^24.0.0",
     "@keystone-next/types": "^24.0.0",
     "@keystone-ui/button": "^5.0.0",
