+++ packages/keystone/admin-ui/utils/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "dist/keystone.cjs.js",
+  "module": "dist/keystone.esm.js"
+}
