+++ packages/keystone/package.json
@@ -29,9 +29,9 @@
     "@babel/plugin-transform-modules-commonjs": "^7.15.0",
     "@babel/runtime": "^7.15.3",
     "@emotion/hash": "^0.8.0",
+    "@emotion/weak-memoize": "^0.2.5",
     "@graphql-tools/schema": "^8.1.1",
     "@hapi/iron": "^6.0.0",
-    "@keystone-next/admin-ui-utils": "^5.0.6",
     "@keystone-next/types": "^24.0.0",
     "@keystone-next/utils": "^1.0.4",
     "@keystone-ui/button": "^5.0.0",
@@ -151,6 +151,7 @@
       "admin-ui/context.tsx",
       "admin-ui/router.tsx",
       "admin-ui/image.tsx",
+      "admin-ui/utils/index.ts",
       "fields/index.ts",
       "fields/types/*/views/index.tsx",
       "fields/types/relationship/views/RelationshipSelect.tsx"
