+++ packages/keystone/src/admin-ui/utils/useLazyMetadata.tsx
@@ -1,8 +1,8 @@
 import { GraphQLError } from 'graphql';
 import type { AuthenticatedItem, VisibleLists, CreateViewFieldModes } from '@keystone-next/types';
 import { useMemo } from 'react';
-import { DeepNullable, makeDataGetter } from '@keystone-next/admin-ui-utils';
 import { DocumentNode, useQuery, QueryResult, ServerError, ServerParseError } from '../apollo';
+import { DeepNullable, makeDataGetter } from './dataGetter';
 
 export type { AuthenticatedItem, VisibleLists, CreateViewFieldModes } from '@keystone-next/types';
