+++ packages/keystone/src/fields/types/relationship/views/cards/InlineEdit.tsx
@@ -1,5 +1,11 @@
 /** @jsx jsx */
 
+import { ListMeta } from '@keystone-next/types';
+import { Button } from '@keystone-ui/button';
+import { jsx, Stack } from '@keystone-ui/core';
+import { useToasts } from '@keystone-ui/toast';
+import { useCallback, useState } from 'react';
+import { Tooltip } from '@keystone-ui/tooltip';
 import {
   deserializeValue,
   ItemData,
@@ -8,13 +14,7 @@ import {
   useChangedFieldsAndDataForUpdate,
   makeDataGetter,
   DataGetter,
-} from '@keystone-next/admin-ui-utils';
-import { ListMeta } from '@keystone-next/types';
-import { Button } from '@keystone-ui/button';
-import { jsx, Stack } from '@keystone-ui/core';
-import { useToasts } from '@keystone-ui/toast';
-import { useCallback, useState } from 'react';
-import { Tooltip } from '@keystone-ui/tooltip';
+} from '../../../../../admin-ui/utils';
 import { gql, useMutation } from '../../../../../admin-ui/apollo';
 import { GraphQLErrorNotice } from '../../../../../admin-ui/components';
 import { useFieldsObj } from './useItemState';
