+++ packages/keystone/src/fields/types/relationship/views/cards/useItemState.tsx
@@ -1,6 +1,6 @@
 import { FieldMeta, ListMeta } from '@keystone-next/types';
 import { useCallback, useMemo, useState } from 'react';
-import { DataGetter, makeDataGetter } from '@keystone-next/admin-ui-utils';
+import { DataGetter, makeDataGetter } from '../../../../../admin-ui/utils';
 import { gql, useQuery } from '../../../../../admin-ui/apollo';
 import { controller } from '../index';
